requirements.txt
catboost==1.1.1
joblib==1.1.1
lightgbm==3.3.5
matplotlib==3.5.2
numpy==1.21.5
pandas==1.4.4
Pillow==9.5.0
plotly==5.9.0
requests==2.28.1
scikit_learn==1.2.2
seaborn==0.11.2
streamlit==1.22.0
xgboost==1.7.4
plotly==5.9.0
openpyxl==3.1.2